TURMERIC-1590: Add a profile for enabling repositories.
If a user doesn't add the settings.xml entry for releng then the runtime
can't find the repositories and the turmeric-project pom file. To fix
this we add a profile called ebayrepos that can be enabled to make
those repositories available, and thus allowing the turmeric-project
pom file to be found.   This was done at the request of the sonarsource
project so that they could run statistics on the project.