TURMERIC-1657: replace printExceptions with logging of exceptions.
There are many spots where the tests just print stack traces and this
clutters up the console.  These should be logged to the runtime log file
so they can be tracked and captured there, but only if the debug level
is set to FINE or DEBUG.

This in itself should greatly reduce the amount of noise that occurs
in the builds when tests are run.