TURMERIC-1657: restet logging to INFO.
Added information to the wiki page about controlling logging during tests
and also set the logging levels to INFO.  This should give a good balance
between what is typically needed and the verboseness of the build.
The wiki page contains the necessary information to understand how to
set approrpiate logging levels.

http://www.ebayopensource.org/wiki/display/TURMERICDOC/Building+from+Source+at+the++Command+Line#BuildingfromSourceattheCommandLine-ControllingLogging